+++ lib/go_transitions.dart
@@ -2,4 +2,4 @@ library go_transitions;
 
 export 'src/go_transition.dart' hide GoTransitionScope;
 export 'src/with_extensions.dart';
-export 'src/transitions.dart';
+export 'src/transitions.dart' show GoTransitions;